st_app/requirements.txt
requests==2.28.2
streamlit==1.20.0
altair==4.1.0
tinvest==3.0.5
influxdb-client==1.33.0
alpaca-trade-api
psaw==0.1.0
pmaw==3.0.0
fake-useragent==0.1.11
selenium==4.4.3
psycopg2==2.8.6
Pillow==9.4.0
pandas==1.5.3
python-dotenv==1.0.0